pyproject.toml
[project]
name = "autodocs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Milansuman", email = "milansuman2016@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.6",
    "langchain-groq>=1.1.1",
    "langgraph>=1.0.6",
    "netra-sdk>=0.1.62",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "tree-sitter>=0.25.2",
    "tree-sitter-python>=0.25.0",
]

[project.scripts]
autodocs = "autodocs:main"

[build-system]
requires = ["uv_build>=0.9.25,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "rich>=14.2.0",
]
